_platform-api-stash/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
alembic>=1.12.0
redis>=5.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
itsdangerous>=2.1.0
prometheus-client>=0.19.0
structlog>=23.2.0
httpx>=0.25.0

